+++ src/App.tsx
@@ -3,7 +3,7 @@ import { BrowserRouter, Route, Routes, Navigate } from 'react-router-dom';
 import Home from './pages/Home';
 import Login from './pages/Login';
 import SignUp from './pages/SignUp';
-import LoginComplete from './pages/Login/components/LoginComplete';
+import LoginComplete from './pages/Login/LoginComplete';
 import TermsAgreement from './pages/TermsAgreement';
 
 import Profile from './pages/Profile';
