+++ src/pages/Login/LoginComplete/index.tsx
@@ -1,12 +1,12 @@
 import React, { useEffect, useState } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
 
-import Loading from '../../../components/Loading';
-import Modal from '../../../components/Modal';
+import Loading from '@/components/Loading';
+import Modal from '@/components/Modal';
 
-import { getUserInfoByJwtApi } from '../../../apis/auth';
-import { handleError } from '../../../apis/util/handleError';
-import { postTermsAgreementApi } from '../../../apis/user';
+import { getUserInfoByJwtApi } from '@/apis/auth';
+import { handleError } from '@/apis/util/handleError';
+import { postTermsAgreementApi } from '@/apis/user';
 
 const LoginComplete: React.FC = () => {
 	const location = useLocation();
